# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark one in-app notification read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/{id}/retry:
    post:
      operationId: retryNotification
      tags: [notifications]
      summary: Retry a failed notification delivery (staff only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/{id}/approve:
    post:
      operationId: approveNotification
      tags: [notifications]
      summary: Approve a customer-facing notification so it may deliver (staff only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notification-templates:
    get:
      operationId: listNotificationTemplates
      tags: [notifications]
      summary: List notification templates (staff only)
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationTemplate"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notification-templates/preview:
    post:
      operationId: previewNotificationTemplate
      tags: [notifications]
      summary: Render a template with sample context (staff only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreviewInput"
      responses:
        "200":
          description: Rendered preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreviewResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notification-templates/test-send:
    post:
      operationId: testSendNotification
      tags: [notifications]
      summary: Send a test in-app notification of a template to yourself (staff only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationTestSendInput"
      responses:
        "201":
          description: Test notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notification-preferences:
    get:
      operationId: listNotificationPreferences
      tags: [notifications]
      summary: List the current user's notification preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationPreference"
    put:
      operationId: setNotificationPreference
      tags: [notifications]
      summary: Set (upsert) one of the current user's notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferenceInput"
      responses:
        "200":
          description: Preference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreference"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    NotificationStatusEvent:
      type: object
      required: [at, status, detail]
      properties:
        at: { type: string, format: date-time }
        status: { type: string }
        detail: { type: string }
    Notification:
      type: object
      required: [id, tenantId, eventType, channel, recipientType, body, status, requiresApproval, attempts, maxAttempts, statusHistory, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        eventType: { type: string }
        channel: { type: string }
        templateId: { type: ["string", "null"] }
        recipientType: { type: string }
        recipientUserId: { type: ["string", "null"] }
        recipientCustomerId: { type: ["string", "null"] }
        recipientAddress: { type: ["string", "null"] }
        subject: { type: ["string", "null"] }
        body: { type: string }
        status: { type: string }
        requiresApproval: { type: boolean }
        approvedByUserId: { type: ["string", "null"] }
        approvedAt: { type: ["string", "null"], format: date-time }
        attempts: { type: integer }
        maxAttempts: { type: integer }
        lastError: { type: ["string", "null"] }
        nextAttemptAt: { type: ["string", "null"], format: date-time }
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/NotificationStatusEvent"
        relatedEntityType: { type: ["string", "null"] }
        relatedEntityId: { type: ["string", "null"] }
        readAt: { type: ["string", "null"], format: date-time }
        sentAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    NotificationTemplate:
      type: object
      required: [id, tenantId, eventType, channel, name, body, customerFacing, enabled, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        eventType: { type: string }
        channel: { type: string }
        name: { type: string }
        subject: { type: ["string", "null"] }
        body: { type: string }
        customerFacing: { type: boolean }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NotificationPreviewInput:
      type: object
      required: [templateId]
      properties:
        templateId: { type: string, minLength: 1 }
        context:
          type: object
          additionalProperties: { type: string }
    NotificationPreviewResult:
      type: object
      required: [channel, body]
      properties:
        channel: { type: string }
        subject: { type: ["string", "null"] }
        body: { type: string }
    NotificationTestSendInput:
      type: object
      required: [templateId]
      properties:
        templateId: { type: string, minLength: 1 }
    NotificationPreference:
      type: object
      required: [id, tenantId, scope, eventType, channel, enabled, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        scope: { type: string }
        userId: { type: ["string", "null"] }
        customerId: { type: ["string", "null"] }
        eventType: { type: string }
        channel: { type: string }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NotificationPreferenceInput:
      type: object
      required: [eventType, channel, enabled]
      properties:
        eventType: { type: string, minLength: 1 }
        channel:
          type: string
          enum: [InApp, Email, SMS, Push]
        enabled: { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...