# lib/api-zod/src/generated/types/notificationPreference.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationPreference {
  id: string;
  tenantId: string;
  scope: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  customerId?: string | null;
  eventType: string;
  channel: string;
  enabled: boolean;
  createdAt: Date;
  updatedAt?: Date;
}