# lib/api-zod/src/generated/types/notificationPreviewInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationPreviewInputContext } from './notificationPreviewInputContext';

export interface NotificationPreviewInput {
  /** @minLength 1 */
  templateId: string;
  context?: NotificationPreviewInputContext;
}